msl_handler divided into two classes
